requirements.txt
amqp==5.1.1
asgiref==3.7.2
Babel==2.12.1
billiard==4.1.0
celery==5.3.1
celery-flower==1.0.1
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
coreapi==2.3.3
coreschema==0.0.4
Django==4.2.4
django-ajax-selects==2.2.0
django-environ==0.11.2
django-filter==23.2
django-rest-framework==0.1.0
django-rest-swagger==2.2.0
djangorestframework==3.14.0
idna==3.4
itypes==1.2.0
Jinja2==3.1.2
kombu==5.3.2
MarkupSafe==2.1.3
openapi-codec==1.3.2
prompt-toolkit==3.0.39
psycopg2==2.9.7
python-dateutil==2.8.2
pytz==2023.3
requests==2.31.0
simplejson==3.19.1
six==1.16.0
sqlparse==0.4.4
tornado==6.3.3
tzdata==2023.3
uritemplate==4.1.1
urllib3==2.0.4
vine==5.0.0
wcwidth==0.2.6
